Extracted has_one_remote to standalone module and refatored ActiveRecord::Relation
